Wire up Delegate Bytes equals and hashcode
* wire up equals and hashCode directly to the delegate
* refactor the delegate classes to reduce code duplication

In Besu this lack of delegation accounts for ~20% of bonsai import time.
